safe delete

+++ src/brand/brand.service.ts
@@ -1,6 +1,7 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
+import { CampaignService } from 'src/campaign/services/campaign.service';
 import { BrandDto } from './dto/brand.dto';
 import { Brand, BrandDocument } from './schemas/brand.schema';
 
@@ -9,6 +10,7 @@ export class BrandService {
   constructor(
     @InjectModel(Brand.name)
     private readonly brandModel: Model<BrandDocument>,
+    private readonly campaignService: CampaignService,
   ) {}
 
   getAllBrands() {
@@ -23,11 +25,29 @@ export class BrandService {
         { new: true },
       );
     } else {
-      return this.brandModel.create({ title: brand.title, client: brand.client, img: brand.img });
+      return this.brandModel.create({
+        title: brand.title,
+        client: brand.client,
+        img: brand.img,
+      });
     }
   }
 
-  removeBrand(brandId: string) {
+  async removeBrand(brandId: string) {
+    const camp = await this.campaignService.findByBrand(brandId);
+    if (Boolean(camp.length)) {
+      throw new HttpException(
+        {
+          status: HttpStatus.FORBIDDEN,
+          error: 'This brand is in use.',
+        },
+        HttpStatus.FORBIDDEN,
+      );
+    }
     return this.brandModel.findByIdAndRemove(brandId);
+  }
+
+  getByClient(clientId: string) {
+    return this.brandModel.find({ client: clientId }).exec();
   }
 }
